+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - use default dhcp option for public network
 - add comments to logging resource configuration
 - use default security list for public network
+- use default routing table for public network
 
 ### Fixed